[Enhancement] Fix some type conversion warning (#12294)

+++ be/src/common/logconfig.cpp
@@ -38,11 +38,11 @@ static bool logging_initialized = false;
 static std::mutex logging_mutex;
 
 static bool iequals(const std::string& a, const std::string& b) {
-    unsigned int sz = a.size();
+    size_t sz = a.size();
     if (b.size() != sz) {
         return false;
     }
-    for (unsigned int i = 0; i < sz; ++i) {
+    for (size_t i = 0; i < sz; ++i) {
         if (tolower(a[i]) != tolower(b[i])) {
             return false;
         }
@@ -206,11 +206,11 @@ void shutdown_logging() {
 
 std::string FormatTimestampForLog(MicrosecondsInt64 micros_since_epoch) {
     time_t secs_since_epoch = micros_since_epoch / 1000000;
-    int usecs = micros_since_epoch % 1000000;
+    int64_t usecs = micros_since_epoch % 1000000;
     struct tm tm_time;
     localtime_r(&secs_since_epoch, &tm_time);
 
-    return StringPrintf("%02d%02d %02d:%02d:%02d.%06d", 1 + tm_time.tm_mon, tm_time.tm_mday, tm_time.tm_hour,
+    return StringPrintf("%02d%02d %02d:%02d:%02d.%06ld", 1 + tm_time.tm_mon, tm_time.tm_mday, tm_time.tm_hour,
                         tm_time.tm_min, tm_time.tm_sec, usecs);
 }
